arm64: dts: a70x0: disable ihb1 axim
commit  1b61c55dc5ca80630fc1bfa7be708c9ccc3ec67c from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

This monitor is connected to the AXI bus of IHB1,
which is not connected in A7040 package.
Accessing this monitor when there’s no clock on the AXI bus,
will cause the access to stuck.
This patch disable the ihb1 axim in order to avoid the stuck.

Change-Id: Id063e9960cc09a3315899f3d1e6536f0b30540c9
Signed-off-by: David Sniatkiwicz <davidsn@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/43202
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>