dma: xilinx: Update to Framebuffer Driver to support dual addr pointers
The 2017.3 version of the Video Framebuffer supports a separate address
pointer for the chroma plane.  This is needed when the chroma plane
is not contiguous with the luma plane for semi-planar formats.  This
patch updates the client API as well.  Additionally, the IP can be
configured for either 32-bit or 64-bit DMA address pointers.  A
new device tree property is added which is used to indicate the address
width and a callback is set during probe to write to memory using
either 32-bit or 64-bit address formats depending on this dts property
as well as the size of dma address space supported on the host.

Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 49026f28a70e778097c131d848c3bf8a2e44ad8d)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>