MGS-3251 [#imx-701] fix power-off issue for ctrl-C exit
there are lots of events to pipeline for database destroy,
wait for event queue empty before do power-off for last process.

side note for the proposed solution:
it is not the way to flush pipeline in drv_release thread,
since the clock may be turned off by other threads unexpectedly,
also the potential deadlock if lock up the power in drv_release.

Date: Aug 31, 2017
Signed-off-by: Xianzhong <xianzhong.li@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>