MLK-17491-7 pinctrl: imx: remove unnecessary fsl,mux_mask property
Remove unnecessary fsl,mux_mask property which is also not documented in
binding doc. As we already have imx_pinctrl_soc_info structure which
represents the SoC specific properties, encode in it instead.

The patch also simplies the code a bit by removing the mux_shift
calculation code which is not necessary as well.

Reviewed-by: Fugang Duan <fugang.duan@nxp.com>
Signed-off-by: Dong Aisheng <aisheng.dong@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>