drm: xilinx: dp: Skip the aux communication when there's no sink
The DRM master may request the aux transaction, for example, power off
request, even when the sink is not connected. This would result in
the timeout. The connection status is stored in the HPD handler
and checked before the aux transaction.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit c335229b70ea41d560c1842ff7138153916b2b27)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>