edac: synopsys: fix incorrect macro defines
device_config[31:30] in MSTR register of ddrc Indicates the configuration
of the device used in the system as follows
- 00 - x4 device
- 01 - x8 device
- 10 - x16 device
- 11 - x32 device
for x16 and x32, instead of 2 and 3 we added 0x10 and 0x11 which is wrong.
This patch corrects these macro defines.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit a5aa440aa346ae898994f288dc67396485ac61f2)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>