crypto: inside-secure: remove unused enqueue return code
commit  3e1a4d572e0c0cf22cb8305b8bc9916dec8ea8f3 from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

When invalidation result is handled, it enqueues the original
request which was the cause for the invalidation, so since
this invalidation result does not require completion, the
enqueue return code will not be used and thus can be dropped.

In any case the request is in the queue regardless to the
enqueue return code.

Change-Id: I2446f49d4aab97303220169bea74cc13f754b75d
Signed-off-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/45613
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/51649
Signed-off-by: Meng Li <Meng.Li@windriver.com>