staging: fsl-dpaa2/eth: Don't set skb->needed_headroom
The latest changes in the buffer layout code try to make sure
we don't realloc skbs forwarded between two DPAA2 Eth interfaces.
The code works, but it's somewhat frail as it relies on network
stack internal formulas.

Instead, it's safer and easier to just not request any guarantee
from the stack. We already double check in the driver the required
headroom size of egress frames and realloc the skb if needed, so
we don't need to add any extra code.

On forwarding between two of our own interfaces, there is no
functional change; for traffic generated on the core, skb realloc
operations are moved from the stack to our driver, with no visible
impact on performance.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>
[Xulin: Original patch taken from NXP lx2160a-early-access-bsp0.3.]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>