MLK-16919-1 ARM: dts: add a new property to indicate mipi mode for csi
in previous version, the property 'csi-mux-mipi' is used to set GPR for MIPI
mode, but on some SoC like iMX8MQ (reuse the CSI) there's no GPR MUX setting
for it.
Add a new property 'fsl,mipi-mode' and check this first for the mode.
If it's MIPI mode, then parse 'csi-mux-mipi'.

Signed-off-by: Robby Cai <robby.cai@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>