MLK-16919-9 dts: add MIPI CSI2 support on iMX8MQ
on EVK board, the camera on MIPI CSI1 and CSI2 port shares the same I2C bus
and the camera has the same slave address. So it's not possible to enable
two ports at same time. The first port is enabled by default. Here to create
a separate dts file to enable second port.

Signed-off-by: Robby Cai <robby.cai@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>