arm*: disable NEON in kernel mode
commit b3a776555e0d465df138d254d6dc3ac1b718ac6d from linux-stable-rt

NEON in kernel mode is used by the crypto algorithms and raid6 code.
While the raid6 code looks okay, the crypto algorithms do not: NEON
is enabled on first invocation and may allocate/free/map memory before
the NEON mode is disabled again.
This needs to be changed until it can be enabled.
On ARM NEON in kernel mode can be simply disabled. on ARM64 it needs to
stay on due to possible EFI callbacks so here I disable each algorithm.

Cc: stable-rt@vger.kernel.org
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
[OP: backport of b3a776555e0d from v4.14-rt branch of linux-stable-rt]
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>