crypto: inside-secure - the context ipad/opad should use the state sz
commit  83baad973c82c38c9933ae1637126f6c5b5b21b0 from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

This patch replaces use of algorithms digest size with the request
state size to copy the ipad and opad in the context. This doesn't fix
anything as the state and digest size are the same for many algorithms,
and for all the hmac currently supported by this driver. However
hmac(sha224) use the sha224 hash function which has a different digest
and state size. This commit prepares the addition of such algorithms.

Change-Id: I70bb0cb449de308737c4f67c8d5d49a9706703a6
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>
Signed-off-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/51537
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/51685
Signed-off-by: Meng Li <Meng.Li@windriver.com>