usb: host: xhci: add support for Light Reset for DWC3 DRD
commit  2cdd9ba555bc829c4885d649940ae94d1d82e8e1 from
https://github.com/Xilinx/linux-xlnx.git

This patch adds support for Light Reset in Host Driver when
the core is in DRD mode.

There is a requirement of LRESET instead of RESET (Hard) in DWC3
controller when it is switching to the peripheral mode.

Signed-off-by: Manish Narani <mnarani@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>