driver: fsl-mc: adjust initialization logic of mc-bus driver
When run command “restool dpni create”, there is a call trace
as below:
[<ffff0000085cc1b0>] iommu_group_get_for_dev+0x60/0x190
[<ffff0000085d2ec8>] arm_smmu_add_device+0x178/0x5e0
[<ffff0000085cb0a8>] iommu_bus_notifier+0xd0/0x128
[<ffff0000080c2294>] notifier_call_chain+0x5c/0xa0
[<ffff0000080c2d88>] blocking_notifier_call_chain+0x68/0x88
[<ffff0000085dd94c>] device_add+0x38c/0x5a8
[<ffff000008937e68>] fsl_mc_device_add+0x200/0x530
[<ffff00000893a318>] dprc_scan_objects+0x270/0x360
[<ffff00000893a520>] dprc_irq0_handler_thread+0x118/0x180
[<ffff0000080eeab0>] irq_thread_fn+0x30/0x70
[<ffff0000080eed84>] irq_thread+0x114/0x1a8
[<ffff0000080c08c8>] kthread+0x138/0x140
[<ffff0000080833b0>] ret_from_fork+0x10/0x20
Code: d63f0020 aa0003f3 b140041f 54fffe68 (f9406400)

In kernel 4.12, arm-smmu is initialized after mc-bus driver
because of upstream commit f6810c15cf97(“iommu/arm-smmu:
Clean up early-probing workarounds”). In this way, there is
not any action to be processed in arm-smmu when mc-bus driver
sends notification during adding devices of MC(management complex).
And in later, it is also failed to set dma_ops of device.

Therefore, in order to make sure mc-bus is initialized normally
in kernel 4.12, it is need to do below adjustment:
- In file fsl-mc-bus.c, keep other initialization priority unchanged,
  only initialize mc-bus driver with module_init so that it is
  initialized behind arm-smmu.
- In file fsl-mc-bus.c, function fsl_mc_device_add(), move function
  fsl_mc_dma_configure() behind device_add() because of upstream
  commit b913efe78a7c(“arm64: dma-mapping: Remove the notifier
  trick to handle early setting of dma_ops”).
- pick fsl_mc_iommu_configure() out of fsl_mc_dma_configure() and
  keep it before device_add(), because it is need to make sure
  that dev->iommu_fwspec is initialized with arm_smmu_ops before
  device_add().

Signed-off-by: Meng Li <Meng.Li@windriver.com>