pinctrl: allow multiple pin controllers for a GPIO
commit 5b7d40d57a9a1d1922c57c14058f521bc39e4c9e from
https://github.com/emutex/ubilinux-kernel.git

Add support for chaining pinctrl drivers controlling the same GPIO line.
This allows for hardware configurations where multiple controllers must
to be configured to change the function or electrical properties of the
line.

Such is the case of the UP board family. On these boards, GPIO and other
SoC functions are provided by a 40-pin Raspberry Pi compatible header,
plus custom headers on certain models. However, for added flexibility,
there is a FPGA between the SoC and these physical pins. In its default
configuration, this FPGA implements a pin controller for FPGA-connected
pins. This setup requires coordination between the SoC GPIO controller
and the FPGA-based controller, hence this patch.

Suggested-by: Dan O'Donovan <dan@emutex.com>
Signed-off-by: Javier Arteaga <javier@emutex.com>
[Yongxin: Add variable is_up_board to limit UP-specific code only for
UP boards and keep the original code unchanged for other boards.]
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>