nxp-imx6: use correct power domain for pd_pci node
In upstream commit 1f1b7ec6714e ("ARM: dts: imx6sx: Add support for PCI power domain"),
the property "power-supply" of node "pd_pci" is set to "reg_pcie". But in freescale's
SDK, the node which is used for power-supply is named as "reg_pcie_phy", so change it
from "reg_pcie" to "reg_pcie_phy".

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>