ARM: OMAP2+: Use pdata-quirks for sgx deassert_hardreset
This commit comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

Use pdata_quirks to provide platform data to the sgx driver.  This is
used to provide a function pointer for the sgx driver to access
omap_device_deassert_hardreset along with the reset name as defined in
the corresponding hwmod entry.

This platform data will not be required when a separate reset driver is
available allowing decoupling from omap_hwmod and omap_device.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Eric Ruei <e-ruei1@ti.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>