xfs: remove XFS_WANT_CORRUPTED_RETURN from dir3 data verifiers
commit 9101d3707b9acae8bbb0d82d47e99cf5c60b3ee5 upstream

Since __xfs_dir3_data_check verifies on-disk metadata, we can't have it
noisily blowing asserts and hanging the system on corrupt data coming in
off the disk.  Instead, have it return a boolean like all the other
checker functions, and only have it noisily fail if we fail in debug
mode.

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Li Wang <li.wang@windriver.com>