bus: fsl-mc: add fsl-mc userspace support
Adding userspace support for the MC (Management Complex)
means exporting an ioctl capable device file representing
the root resource container.
This new functionality in the fsl-mc bus driver intends to
provide userspace applications an interface to interact with the MC firmware.
Commands that are composed in userspace are sent to the MC firmware
through the RESTOOL_SEND_MC_COMMAND ioctl.
By default the implicit MC I/O portal is used for this operation,
but if the implicit one is busy, a dynamic portal is allocated and then
freed upon execution.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
[Xulin: Original patch taken from NXP lx2160a-early-access-bsp0.3.]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>