arm64: dts: stratix10: use clock bindings for the Stratix10 platform
Reverts 56594bb5408396a53a0794f22a31ea98fc999299
(FogBugz #488843-1: dts: clock bindings for the SoCFPGA Stratix10)

and

Backports
commit d93101abe41e9596555a0f7f6f775e543b71c441 upstream

Use the clock bindings for the Stratix10 SoC. This includes changing the
old binding of "intc,clk-s10-mgr" to "intel,stratix10-clkmgr". The reason that
this can be done is that there are currently no clock driver for Stratix10,
thus there are no consumers of the old binding. So changing the binding will
not break any legacy code.

Signed-off-by: Dinh Nguyen <dinguyen@kernel.org>
Signed-off-by: Marian Florea <marian.florea@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>