bman:add kexec & kdump support
When booting second kernel in kexec, need check device state when initial
device.

In the kdump boot, there are still buffers in the bman buffer pools
released by the first kernel. These buffers have nothing to do with
the current kernel, and may cause issues when we try to use these
buffers. So we should drain all the buffer pools before kdump boot.

Signed-off-by: Jiang Lu <lu.jiang@windriver.com>