nxp-imx8: add MU_SendMessage timeout
The rpmsg is a virtio-based messaging bus which allows kernel
drivers to communicate with remote processors available on
the system. The MU (Message Unit) in imx8 is used to implement
this function. The A side of MU is Cortex A53 and the B side of
MU is Cortex M4. But the Cortex M4 is in RESET mode without running
FreeRTOS on it. When imx_rpmsg_notify calls MU_SendMessage to send
a message to Cortex M4, the message will stay in MUA_ATRx register
and M4 will not read it in MUB_ARRx register. This means that
once imx_rpmsg_notify is called, the TEn(Processor A Transmit Register n Empty)
bit in MUA_ASR will stay zero. When the kexec kernel boots, MU_SendMessage
will run in a indefinite loop because the first boot kernel send message
to M4 failed.

So change indefinite loop to a time limited delay to avoid this.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>